requirements.txt
fastapi
uvicorn
google-cloud-bigquery
pydantic
openai